Refactor DB manager and improve connectivity
